pyproject.toml
[project]
name = "vcon-server"
version = "0.1.0"
requires-python = ">=3.12,<3.14"

# Core dependencies — shared by both the api and conserver services.
dependencies = [
    "opentelemetry-api>=1.38.0",
    "opentelemetry-sdk>=1.38.0",
    "opentelemetry-exporter-otlp>=1.38.0",
    "opentelemetry-distro==0.59b0",
    "redis==4.6.0",
    "pyyaml>=6.0.1",
    "sentry-sdk>=2.8.0",
    "pydantic>=2.6.2",
    "tenacity>=8.2.3",
    "uuid6>=2024.1.12",
    "pydash>=7.0.7",
    "requests>=2.31.0",
    "pycose>=1.0.1",
    "ecdsa>=0.18.0",
    "python-dotenv>=1.0.1",
    "python-json-logger==2.0.4",
    "python-dateutil>=2.8.2",
    # Canonical vCon construction/serialization library.
    # 0.9.2 emits the `vcon: "0.4.0"` syntax param and drops empty
    # `group`/`redacted` defaults — see vcon-speckit non-negotiables.
    "vcon>=0.9.2",
]

[dependency-groups]
# Storage backend packages — needed by both api (reads) and conserver (writes).
# Install with: uv sync --group storage
storage = [
    "pymongo>=4.7.2",
    "elasticsearch>=8.13.1,<9.0.0",
    "pymilvus>=2.3.0",
    "msal>=1.32.3",
    "boto3>=1.34.52",
    "psycopg2-binary>=2.9.9",
    "peewee>=3.17.1",
]

# API service dependencies — HTTP layer and API key management.
# Install with: uv sync --group api
api = [
    {include-group = "storage"},
    "fastapi==0.115.6",
    "uvicorn==0.23.2",
    "gunicorn>=23.0.0",
    "starlette>=0.40.0",
]

# Conserver service dependencies — processing links (ML/AI/audio) and storage backends.
# Install with: uv sync --group conserver
conserver = [
    {include-group = "storage"},
    # Processing links
    "transformers>=4.48.0",
    "openai>=1.60.0",
    "groq>=0.4.0",
    "deepgram-sdk>=3.1.5,<4.0.0",
    "slack-sdk>=3.27.1",
    "ffmpeg-python>=0.2.0",
    "pydub>=0.25.1",
    "jq>=1.8.0",
    # Hot-reload wrapper used by docker-compose watchmedo
    "watchdog",
]

# Development / test dependencies.
dev = [
    "black>=24.2.0",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.23.5",
    "pytest-dotenv>=0.5.2",
    "pytest-redis>=3.1.3",
    "faker>=33.3.1",
    "trio>=0.28.0",
    "anyio>=4.8.0",
    "httpx>=0.27.0",
    "pytest-cov>=7.1.0",
]


[tool.black]
line-length = 120
skip-string-normalization = true

[tool.coverage.run]
branch = true
source = ["api", "common", "conserver"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/conftest.py",
]

[tool.coverage.report]
show_missing = true
skip_covered = true
